v0.2 update
Changed method of removing/replacing body background. This removed any wallpapers in a more reliable and universally applicable way (hopefully) and replaces the background with a solid color (#222426)
Added a style override to remove the Featured Video element that appears on some pages. I had overlooked this as in my previous testing I had not gone to a large enough page that warranted a video.