Tidying Up
I've broken down the style sheet into sections for each main page as well as a section at the top for style adjustments that span multiple pages. Also update the font to the correct OG Blaseball font, Lora, from Google Fonts.

Lora is currently paved over every other font type on the page but this is just a placeholder for now.