v1.4.1 Game Field Size Increase
First pass at increasing the size of the game field indicator (1st, 2nd, 3rd base). There's not much room to work with so I can't make them much larger than this without adding in some more styling to create or free up space elsewhere, but for now this is fine.